Breaking down files into atomic design
