+++ src/client/components/Atoms/ShoppingListIcon/XIcon.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import XIcon from './Xicon.svg';
+
+const XIconComponent = () => <img src={XIcon} alt="Remove" />;
+
+export default XIconComponent;
